refactor: delete old jobs code
- deleted engine/jobs.py, engine/azure_helpers.py,
  engine/azure_engine.py
- CharacterClass moved to engine/data_loader.py
- getLowerWeaponRanks moved to azure_constants
- get_stat_modifier removed
- character.py uses CLASS_DEFINITIONS instead of CREATION_RULES
