+++ tests/discord_integration/test_session.py
@@ -9,8 +9,7 @@ the channel. No human interaction required.
 import pytest
 
 from engine import create_character, open_turn, start_session
-from engine.azure_engine import CharacterClass
-from models import SessionMode, TurnStatus
+from models import CharacterClass, SessionMode, TurnStatus
 from tests.discord_integration._config import TEST_CHANNEL_ID, TEST_DM_USER_ID
