+++ tests/test_combat_engine.py
@@ -32,11 +32,11 @@ from engine import (
     start_session,
     submit_turn,
 )
-from engine.azure_engine import CharacterClass
 from engine.combat import CombatAction, _npc_decide, _tick_conditions
 from models import (
     NPC,
     ActiveCondition,
+    CharacterClass,
     GameState,
     Party,
     RangeBand,
@@ -1336,9 +1336,9 @@ class TestWeaponIdSelection:
         """Set up combat state with a mage holding a spellbook (multiple weapons)."""
         from engine import create_character, equip_item, give_item
         from engine.azure_constants import ItemSlot
-        from engine.character import CharacterClass
         from engine.data_loader import ITEM_REGISTRY
         from engine.item import ChargeWeapon, ContainerItem
+        from models import CharacterClass
 
         state = _make_state_with_npc()
