+++ tests/test_two_handed_tag.py
@@ -28,8 +28,7 @@ sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
 from engine import create_character, start_session
 from engine.azure_constants import ItemSlot
-from engine.azure_engine import CharacterClass
-from models import GameState, InventoryItem, Party
+from models import CharacterClass, GameState, InventoryItem, Party
 
 
 def _make_state_with_char():
